Switch away from the AppImage-based LOVE
It doesn't work well in a CI environment. Because FUSE isn't supported
in non-priveleged Docker containers, you have to extract the AppImage,
which takes away some of the advantage when compared to a .tar.gz
archive.

More importantly, the AppImage version of LOVE always tries to start the
SDL video subsystem, regardless of the love.window and love.graphics
modules being turned off.